feat: filter manager applications
feat: extract common methods for project fetching from registry

feat: extract common methods for project fetching from registry

feat: scaffold application filtering
