+++ packages/builder/src/actions/grantsMetadata.ts
@@ -1,12 +1,13 @@
 import { datadogRum } from "@datadog/browser-rum";
 import { ethers } from "ethers";
 import { Dispatch } from "redux";
+import { getProviderByChainId } from "common";
 import { RootState } from "../reducers";
 import ProjectRegistryABI from "../contracts/abis/ProjectRegistry.json";
 import PinataClient from "../services/pinata";
 import { LocalStorage } from "../services/Storage";
 import { Metadata, ProjectRegistryMetadata } from "../types";
-import { getProjectURIComponents, getProviderByChainId } from "../utils/utils";
+import { getProjectURIComponents } from "../utils/utils";
 
 export const GRANT_METADATA_LOADING_URI = "GRANT_METADATA_LOADING_URI";
 export interface GrantMetadataLoadingURI {
@@ -197,7 +198,9 @@ export const fetchGrantData =
 
     const chainID = Number(chainId);
     const addresses = { projectRegistry: registryAddress };
-    const appProvider = getProviderByChainId(chainID);
+    const appProvider = getProviderByChainId(
+      chainID
+    ) as unknown as ethers.providers.BaseProvider;
 
     let project: ProjectRegistryMetadata;
