+++ packages/builder/src/utils/utils.ts
@@ -61,23 +61,6 @@ export const getProjectURIComponents = (id: string) => {
   };
 };
 
-export const getProviderByChainId = (chainId: number) => {
-  const { web3Provider } = global;
-
-  const chainConfig = web3Provider?.chains?.find(
-    // Yes, parameter type for chainId is number, but sometimes we pass it as a string
-    // so adding a cast to Number just in case
-    (i) => i.id === Number(chainId)
-  );
-
-  if (!chainConfig) {
-    throw new Error(`chainConfig not found for chain ID ${chainId}`);
-  }
-
-  // TODO: Create a more robust RPC here to avoid fails
-  return ethers.getDefaultProvider(chainConfig.rpcUrls.default);
-};
-
 export const getAddressType = async (address: string): Promise<AddressType> => {
   const { web3Provider } = global;
