+++ packages/common/src/index.ts
@@ -1,6 +1,9 @@
 import useSWR from "swr";
 import { useParams } from "react-router";
 import { isAddress } from "viem";
+import { ethers } from "ethers";
+import ProjectRegistryABI from "./abis/ProjectRegistry.json";
+import { global } from "./global";
 
 export enum ChainId {
   MAINNET = 1,
@@ -235,9 +238,7 @@ export function fetchProjectPaidInARound(
     }
   );
 
-  const payouts = data?.data?.payoutStrategies[0]?.payouts || [];
-
-  return payouts;
+  return data?.data?.payoutStrategies[0]?.payouts || [];
 }
 
 /** Returns the current round id extracted from the current  route
@@ -319,4 +320,89 @@ export const convertStatusToText = (
     default:
       return "PENDING";
   }
+};
+
+/**
+ * Verifies that the submitter of an application is also the project owner.
+ * */
+export function verifyOwnerOfApplication(
+  chainId: ChainId,
+  projectId: string,
+  sender: string
+): boolean {
+  /* Get current owners of a project */
+  /*TODO: what to do about applications of past owners? */
+  const owners = fetchProjectOwners(chainId, projectId);
+
+  /* Verify that sender is in the owners of the project*/
+  return owners.includes(sender);
+}
+
+export const fetchProjectOwners = (chainID: number, projectID: string) => {
+  const addresses = addressesByChainID(chainID);
+  const appProvider = getProviderByChainId(chainID);
+
+  const projectRegistry = new ethers.Contract(
+    addresses.projectRegistry!,
+    ProjectRegistryABI,
+    appProvider
+  );
+
+  return projectRegistry.getProjectOwners(projectID) as string[];
+};
+
+export const getProviderByChainId = (chainId: number) => {
+  const { web3Provider } = global;
+
+  const chainConfig = web3Provider?.chains?.find(
+    // Yes, parameter type for chainId is number, but sometimes we pass it as a string
+    // so adding a cast to Number just in case
+    (i) => i.id === Number(chainId)
+  );
+
+  if (!chainConfig) {
+    throw new Error(`chainConfig not found for chain ID ${chainId}`);
+  }
+
+  // TODO: Create a more robust RPC here to avoid fails
+  return ethers.getDefaultProvider(chainConfig.rpcUrls.default);
+};
+
+export const addressesByChainID = (chainID: number) => {
+  const chainName: string = chains[chainID];
+  return addresses[chainName];
+};
+export const chains: { [key: number]: string } = {
+  31337: "localhost",
+  5: "goerli",
+  10: "optimism",
+  69: "optimisticKovan", // todo: update to 420: "optimisticGoerli"
+  4002: "fantomTestnet",
+  250: "fantom",
+  1: "mainnet",
+};
+export const addresses: DeploymentAddress = {
+  localhost: {
+    projectRegistry: "0x832c5391dc7931312CbdBc1046669c9c3A4A28d5",
+  },
+  optimism: {
+    projectRegistry: "0x8e1bD5Da87C14dd8e08F7ecc2aBf9D1d558ea174",
+  },
+  mainnet: {
+    projectRegistry: "0x03506eD3f57892C85DB20C36846e9c808aFe9ef4",
+  },
+  goerli: {
+    projectRegistry: "0xa71864fAd36439C50924359ECfF23Bb185FFDf21",
+  },
+  fantomTestnet: {
+    projectRegistry: "0x984749e408FF0446d8ADaf20E293F2F299396631",
+  },
+  fantom: {
+    projectRegistry: "0x8e1bD5Da87C14dd808F7ecc2aBf9D1d558ea174",
+  },
+};
+export type DeploymentAddress = {
+  [key: string]: {
+    projectRegistry: string | undefined;
+  };
 };
